Move repository drivers up to RegistryManager
Why you made the change:

I did this because it pushed the dependencies up and creates a cleaner
interface across these boundaries. Also it moves the drivers
configuration/construction closer to where that data would actually come
in.
