+++ lib/togls.rb
@@ -9,8 +9,7 @@ require 'togls/feature_repository_drivers/in_memory_driver'
 require 'togls/rule_repository_drivers'
 require 'togls/rule_repository_drivers/in_memory_driver'
 require 'togls/toggler'
-require 'togls/release_toggle_registry'
-require 'togls/test_toggle_registry'
+require 'togls/toggle_registry'
 require 'togls/feature_repository'
 require 'togls/rule_repository'
 require 'togls/toggle_repository'
